+++ src/main/java/cn/pasteme/common/annotation/Timer.java
@@ -1,0 +1,20 @@
+package cn.pasteme.common.annotation;
+
+import org.springframework.core.annotation.Order;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+/**
+ * 记录一个函数的执行时间
+ *
+ * @author Lucien
+ * @version 1.0.0
+ */
+@Target({ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+@Order(2147483647)
+public @interface Timer {
+}
